+++ packages/lib_llama_cpp_macos/lib/lib_llama_cpp_macos.dart
@@ -26,10 +26,7 @@ final class LibLlamaCppMacos extends LibLlamaCppPlatform {
     );
   }
 
-  static const _capabilities = {
-    LlamaCppLibraryCapability.cpu,
-    LlamaCppLibraryCapability.metal,
-  };
+  static const _capabilities = {LlamaCppLibraryCapability.cpu};
 }
 
 Set<LlamaCppLibraryCapability> _capabilitiesForPreferredPath(