+++ packages/lib_llama_cpp_windows/lib/lib_llama_cpp_windows.dart
@@ -26,10 +26,7 @@ final class LibLlamaCppWindows extends LibLlamaCppPlatform {
     );
   }
 
-  static const _capabilities = {
-    LlamaCppLibraryCapability.cpu,
-    LlamaCppLibraryCapability.vulkan,
-  };
+  static const _capabilities = {LlamaCppLibraryCapability.cpu};
 }
 
 Set<LlamaCppLibraryCapability> _capabilitiesForPreferredPath(